+++ tests/04 - basic-templates.test.ts
@@ -1,4 +1,7 @@
-import HomeAssistantJavaScriptTemplates, { HomeAssistantJavaScriptTemplatesRenderer } from '../src';
+import {
+    HomeAssistantJavaScriptTemplates,
+    HomeAssistantJavaScriptTemplatesRenderer
+} from '../src/classes';
 import { HOME_ASSISTANT_ELEMENT, HASS } from './constants';
 
 describe('Basic templates tests', () => {